Change action control method, add reward noise, add ign_partition
Old control method is by modifying the previous controller target. This is now removed in favor of setting controller target as current joint state (with noise) + action.

Reward noise is to maybe cause the agent to exploit less

ign_partition is to fix a bug whereby the marker spawning and modifying are interfering with each other across different simulation instances. This is due to it using ign-transport which doesn't respect GAZEBO_MASTER_URI but IGN_PARTITION.